blogapi/pyproject.toml
[project]
name = "blogapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]
